[core] Test some TypeScript perf optimizations
