+++ packages/mui-material/src/index.d.ts
@@ -21,44 +21,6 @@ export type StandardProps<
     style?: React.CSSProperties;
   };
 
-/**
- * @internal
- * ONLY USE FROM WITHIN mui/material-ui
- *
- * Internal helper type for conform (describeConformance) components
- * However, we don't declare classes on this type.
- * It is recommended to declare them manually with an interface so that each class can have a separate JSDoc.
- */
-export type InternalStandardProps<C, Removals extends keyof C = never> = DistributiveOmit<
-  C,
-  'classes' | Removals
-> &
-  // each component declares it's classes in a separate interface for proper JSDoc
-  StyledComponentProps<never> & {
-    ref?: C extends { ref?: infer RefType } ? RefType : React.Ref<unknown>;
-    // TODO: Remove implicit props. Up to each component.
-    className?: string;
-    style?: React.CSSProperties;
-  };
-
-export type PaletteMode = 'light' | 'dark';
-export interface Color {
-  50: string;
-  100: string;
-  200: string;
-  300: string;
-  400: string;
-  500: string;
-  600: string;
-  700: string;
-  800: string;
-  900: string;
-  A100: string;
-  A200: string;
-  A400: string;
-  A700: string;
-}
-
 export namespace PropTypes {
   // keeping the type structure for backwards compat
   // eslint-disable-next-line @typescript-eslint/no-shadow, @typescript-eslint/no-unused-vars
@@ -73,6 +35,10 @@ export { colors };
 export * from './styles';
 
 export * from './utils';
+
+export { Color, PaletteMode } from './styles/createPalette';
+
+export { InternalStandardProps } from './InternalStandardProps.types';
 
 export { default as Accordion } from './Accordion';
 export * from './Accordion';
