+++ packages/mui-material/src/styles/createTheme.d.ts
@@ -1,11 +1,12 @@
-import { ThemeOptions as SystemThemeOptions, Theme as SystemTheme } from '@mui/system';
-import { Mixins, MixinsOptions } from './createMixins';
+import { ThemeOptions as SystemThemeOptions } from '@mui/system';
+import { MixinsOptions } from './createMixins';
 import { Palette, PaletteOptions } from './createPalette';
-import { Typography, TypographyOptions } from './createTypography';
+import { TypographyOptions } from './createTypography';
 import { Shadows } from './shadows';
-import { Transitions, TransitionsOptions } from './createTransitions';
-import { ZIndex, ZIndexOptions } from './zIndex';
+import { TransitionsOptions } from './createTransitions';
+import { ZIndexOptions } from './zIndex';
 import { Components } from './components';
+import { BaseTheme } from './BaseTheme.types';
 
 export interface ThemeOptions extends Omit<SystemThemeOptions, 'zIndex'> {
   mixins?: MixinsOptions;
@@ -15,16 +16,6 @@ export interface ThemeOptions extends Omit<SystemThemeOptions, 'zIndex'> {
   transitions?: TransitionsOptions;
   typography?: TypographyOptions | ((palette: Palette) => TypographyOptions);
   zIndex?: ZIndexOptions;
-  unstable_strictMode?: boolean;
-}
-
-interface BaseTheme extends SystemTheme {
-  mixins: Mixins;
-  palette: Palette;
-  shadows: Shadows;
-  transitions: Transitions;
-  typography: Typography;
-  zIndex: ZIndex;
   unstable_strictMode?: boolean;
 }
